+++ app/src/main/java/woowacourse/movie/mapper/PeopleCountMapper.kt
@@ -1,7 +1,7 @@
 package woowacourse.movie.mapper
 
 import woowacourse.movie.domain.PeopleCount
-import woowacourse.movie.model.PeopleCountModel
+import woowacourse.movie.uimodel.PeopleCountModel
 
 fun PeopleCount.toModel(): PeopleCountModel = PeopleCountModel(
     count = count,
