+++ app/src/main/java/woowacourse/movie/mapper/SelectedSeatsMapper.kt
@@ -1,7 +1,7 @@
 package woowacourse.movie.mapper
 
 import woowacourse.movie.domain.seat.SelectedSeats
-import woowacourse.movie.model.SelectedSeatsModel
+import woowacourse.movie.uimodel.SelectedSeatsModel
 
 fun SelectedSeats.toModel(): SelectedSeatsModel = SelectedSeatsModel(
     seats = seats.map { it.toModel() }.toSet(),
