+++ app/src/main/java/woowacourse/movie/ui/home/adapter/MovieItemViewHolder.kt
@@ -6,7 +6,7 @@ import android.widget.ImageView
 import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
 import woowacourse.movie.R
-import woowacourse.movie.model.MovieListModel
+import woowacourse.movie.uimodel.MovieListModel
 import java.time.LocalDate
 import java.time.format.DateTimeFormatter
