+++ app/src/main/java/woowacourse/movie/ui/moviedetail/MovieDetailActivity.kt
@@ -14,9 +14,9 @@ import android.widget.Spinner
 import android.widget.TextView
 import androidx.appcompat.app.AppCompatActivity
 import woowacourse.movie.R
-import woowacourse.movie.model.MovieListModel
-import woowacourse.movie.model.PeopleCountModel
 import woowacourse.movie.ui.seat.SeatSelectionActivity
+import woowacourse.movie.uimodel.MovieListModel
+import woowacourse.movie.uimodel.PeopleCountModel
 import woowacourse.movie.utils.failLoadingData
 import woowacourse.movie.utils.getParcelableCompat
 import woowacourse.movie.utils.getSerializableExtraCompat
