+++ app/src/main/java/woowacourse/movie/ui/seat/SeatSelectionActivity.kt
@@ -13,15 +13,15 @@ import woowacourse.movie.domain.seat.Seat
 import woowacourse.movie.domain.seat.SelectedSeats
 import woowacourse.movie.mapper.toDomain
 import woowacourse.movie.mapper.toModel
-import woowacourse.movie.model.MovieTicketModel
-import woowacourse.movie.model.PeopleCountModel
-import woowacourse.movie.model.ReservationModel
-import woowacourse.movie.model.SeatModel
-import woowacourse.movie.model.SelectedSeatsModel
 import woowacourse.movie.ui.alarm.ReservationAlarmManager
 import woowacourse.movie.ui.moviedetail.MovieDetailActivity
 import woowacourse.movie.ui.setting.SettingFragment
 import woowacourse.movie.ui.ticket.MovieTicketActivity
+import woowacourse.movie.uimodel.MovieTicketModel
+import woowacourse.movie.uimodel.PeopleCountModel
+import woowacourse.movie.uimodel.ReservationModel
+import woowacourse.movie.uimodel.SeatModel
+import woowacourse.movie.uimodel.SelectedSeatsModel
 import woowacourse.movie.utils.failLoadingData
 import woowacourse.movie.utils.getPreferences
 import woowacourse.movie.utils.getSerializableExtraCompat
