+++ app/src/main/java/woowacourse/movie/uimodel/MovieTicketModel.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.model
+package woowacourse.movie.uimodel
 
 import java.io.Serializable
 import java.time.LocalDateTime
