+++ app/src/main/java/woowacourse/movie/uimodel/ReservationModel.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.model
+package woowacourse.movie.uimodel
 
 object ReservationModel {
     private val _tickets = mutableListOf<MovieTicketModel>()
