+++ app/src/main/java/woowacourse/movie/uimodel/SeatModel.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.model
+package woowacourse.movie.uimodel
 
 import android.content.Context
 import android.view.LayoutInflater
