+++ app/src/main/java/woowacourse/movie/uimodel/SelectedSeatsModel.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.model
+package woowacourse.movie.uimodel
 
 import java.io.Serializable
